+++ ra2ce/analyses/analysis_config_data/analysis_config_data_reader.py
@@ -34,6 +34,13 @@ from ra2ce.analyses.analysis_config_data.analysis_config_data import (
     IndirectAnalysisNameList,
     ProjectSection,
 )
+from ra2ce.analyses.analysis_config_data.enums.analysis_direct_enum import (
+    AnalysisDirectEnum,
+)
+from ra2ce.analyses.analysis_config_data.enums.analysis_indirect_enum import (
+    AnalysisIndirectEnum,
+)
+from ra2ce.analyses.analysis_config_data.enums.weighing_enum import WeighingEnum
 from ra2ce.common.configuration.ini_configuration_reader_protocol import (
     ConfigDataReaderProtocol,
 )
@@ -96,12 +103,21 @@ class AnalysisConfigDataReader(ConfigDataReaderProtocol):
         self, section_name: str
     ) -> AnalysisSectionIndirect:
         _section = AnalysisSectionIndirect(**self._parser[section_name])
+        _section.analysis = AnalysisIndirectEnum.get_enum(
+            self._parser.get(section_name, "analysis", fallback=None)
+        )
         _section.save_gpkg = self._parser.getboolean(
             section_name, "save_gpkg", fallback=_section.save_gpkg
         )
         _section.save_csv = self._parser.getboolean(
             section_name, "save_csv", fallback=_section.save_csv
         )
+        _weighing = self._parser.get(section_name, "weighing", fallback=None)
+        # Map distance -> length
+        if _weighing == "distance":
+            _section.weighing = WeighingEnum.LENGTH
+        else:
+            _section.weighing = WeighingEnum.get_enum(_weighing)
         # losses
         _section.traffic_cols = self._parser.getlist(
             section_name, "traffic_cols", fallback=_section.traffic_cols
@@ -180,6 +196,9 @@ class AnalysisConfigDataReader(ConfigDataReaderProtocol):
 
     def _get_analysis_section_direct(self, section_name: str) -> AnalysisSectionDirect:
         _section = AnalysisSectionDirect(**self._parser[section_name])
+        _section.analysis = AnalysisDirectEnum.get_enum(
+            self._parser.get(section_name, "analysis", fallback=None)
+        )
         _section.save_gpkg = self._parser.getboolean(
             section_name, "save_gpkg", fallback=_section.save_gpkg
         )