+++ tests/analyses/indirect/test_origin_closest_destination.py
@@ -2,7 +2,9 @@ from ra2ce.analyses.analysis_config_data.analysis_config_data import (
     AnalysisConfigData,
     AnalysisSectionIndirect,
 )
+from ra2ce.analyses.analysis_config_data.enums.weighing_enum import WeighingEnum
 from ra2ce.analyses.indirect.origin_closest_destination import OriginClosestDestination
+from ra2ce.graph.graph_files.graph_files_collection import GraphFilesCollection
 from ra2ce.graph.network_config_data.network_config_data import (
     NetworkSection,
     OriginsDestinationsSection,
@@ -23,12 +25,16 @@ class TestOriginClosestDestination:
             ),
             network=NetworkSection(file_id=""),
         )
-        _analysis = AnalysisSectionIndirect(threshold="", weighing="")
+        _analysis = AnalysisSectionIndirect(threshold="", weighing=WeighingEnum.INVALID)
+        _graph_files = GraphFilesCollection()
         _hazard_names = None
 
         # 2. Run test.
         _ocd = OriginClosestDestination(
-            config=_config_dict, analysis=_analysis, hazard_names_df=_hazard_names
+            config=_config_dict,
+            analysis=_analysis,
+            graph_files=_graph_files,
+            hazard_names_df=_hazard_names,
         )
 
         # 3. Verify expectations.