+++ tests/runners/test_indirect_analysis_runner.py
@@ -1,6 +1,10 @@
 import pytest
+
 from ra2ce.analyses.analysis_config_data.analysis_config_data import (
     AnalysisSectionIndirect,
+)
+from ra2ce.analyses.analysis_config_data.enums.analysis_indirect_enum import (
+    AnalysisIndirectEnum,
 )
 from ra2ce.configuration.config_wrapper import ConfigWrapper
 from ra2ce.runners.indirect_analysis_runner import IndirectAnalysisRunner
@@ -23,7 +27,9 @@ class TestIndirectAnalysisRunner:
     ):
         # 1. Define test data.
         dummy_ra2ce_input.analysis_config.config_data.analyses = [
-            AnalysisSectionIndirect(analysis="single_link_redundancy")
+            AnalysisSectionIndirect(
+                analysis=AnalysisIndirectEnum.SINGLE_LINK_REDUNDANCY
+            )
         ]
 
         # 2. Run test.